requirements.txt
starlette[full]
pydantic
uvicorn
aiofiles
passlib
psycopg2-binary
python-decouple
python-slugify
wtforms
PyJWT
python-magic-bin
Pillow
websockets
databases[sqlite]
databases[postgresql]
gunicorn
asyncpg
whitenoise
json2html
aiocsv
